折线图基本完成了，下面还可以加上滚动条和总在最前等等设定，下面开始处理光电的曲线图

+++ SRB_Frame/support/support.cs
@@ -96,9 +96,9 @@ namespace SRB.Frame
             return a;
         }
 
-        static public double getElapsedMs(this System.Diagnostics.Stopwatch sw)
+        static public float getElapsedMs(this System.Diagnostics.Stopwatch sw)
         {
-            return (1000.0 * sw.ElapsedTicks) / System.Diagnostics.Stopwatch.Frequency; ;
+            return (1000.0f * sw.ElapsedTicks) / System.Diagnostics.Stopwatch.Frequency; ;
         }
